Move sample download to safety page
