+++ src/components/sample/SamplesTable.tsx
@@ -1,11 +1,7 @@
-import { Tooltip, IconButton } from '@material-ui/core';
-import { Visibility } from '@material-ui/icons';
-import GetAppIcon from '@material-ui/icons/GetApp';
-import MaterialTable, { MaterialTableProps } from 'material-table';
+import MaterialTable, { MaterialTableProps, Column } from 'material-table';
 import React from 'react';
 import { DecodedValueMap, SetQuery, QueryParamConfig } from 'use-query-params';
 
-import { useDownloadPDFSample } from 'hooks/sample/useDownloadPDFSample';
 import { SampleBasic } from 'models/Sample';
 import { tableIcons } from 'utils/materialIcons';
 
@@ -14,56 +10,23 @@ type SamplesTableQueryParamsType = {
   search: QueryParamConfig<string | null | undefined>;
 };
 
+const defaultColumns: Column<SampleBasic>[] = [
+  { title: 'Title', field: 'title' },
+  { title: 'Status', field: 'safetyStatus' },
+  { title: 'Created', field: 'created' },
+];
+
 const SamplesTable = (
   props: Omit<MaterialTableProps<SampleBasic>, 'columns'> & {
     urlQueryParams: DecodedValueMap<SamplesTableQueryParamsType>;
     setUrlQueryParams: SetQuery<SamplesTableQueryParamsType>;
-    setSelectedSample: (sample: SampleBasic) => void;
+    columns?: Column<SampleBasic>[];
   }
 ) => {
-  const downloadPDFSample = useDownloadPDFSample();
-  const RowActionButtons = (rowData: SampleBasic) => {
-    const iconButtonStyle = { padding: '7px' };
-
-    return (
-      <>
-        <Tooltip title="View sample">
-          <IconButton
-            style={iconButtonStyle}
-            onClick={() => props.setSelectedSample(rowData)}
-          >
-            <Visibility />
-          </IconButton>
-        </Tooltip>
-        <Tooltip title="Download sample as pdf">
-          <IconButton
-            data-cy="download-sample"
-            onClick={() => downloadPDFSample(rowData.id)}
-            style={iconButtonStyle}
-          >
-            <GetAppIcon />
-          </IconButton>
-        </Tooltip>
-      </>
-    );
-  };
-
-  const columns = [
-    {
-      title: 'Actions',
-      sorting: false,
-      removable: false,
-      render: RowActionButtons,
-    },
-    { title: 'Title', field: 'title' },
-    { title: 'Status', field: 'safetyStatus' },
-    { title: 'Created', field: 'created' },
-  ];
-
   return (
     <MaterialTable
+      columns={props.columns ? props.columns : defaultColumns}
       icons={tableIcons}
-      columns={columns}
       title="Samples"
       onSearchChange={searchText => {
         props.setUrlQueryParams({
@@ -76,16 +39,6 @@ const SamplesTable = (
         selection: true,
       }}
       {...props}
-      actions={[
-        {
-          icon: GetAppIcon,
-          tooltip: 'Download sample',
-          onClick: (event, rowData) =>
-            downloadPDFSample(
-              (rowData as SampleBasic[]).map(({ id }) => id).join(',')
-            ),
-        },
-      ]}
     />
   );
 };
